requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorgram.py==1.2.0
frozenlist==1.4.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==5.1.0
multidict==6.0.4
numpy==1.26.3
pandas==2.1.4
pillow==10.2.0
prettytable==3.9.0
PyJWT==2.8.0
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
twilio==8.12.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
yarl==1.9.4
